+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to Rainbow-CLI will be documented in this file.
 ## [1.53.0] - 2018-03-10
 
 -   RQRAINB-1255: CPAAS / Rainbow CLI / Add Dashboard In deployment
+-   RQRAINB-1257: CPAAS / Rainbow CLI / Manage Oauth commands
 
 ## [1.52.3] - 2018-02-16